프로그레스바 바뀐 타입 수정

+++ design-system/src/main/java/com/danbam/design_system/component/Item.kt
@@ -71,7 +71,7 @@ fun FundingItem(
                 fontSize = 12
             )
             Spacer(modifier = Modifier.weight(1F))
-            IndiStrawProgress(currentProgress = item.percentage.toFloat())
+            IndiStrawProgress(currentProgress = item.percentage)
         }
     }
 }
